requirements.txt
Django
Pillow
psycopg2-binary
djangorestframework
djangorestframework_simplejwt
drf-yasg
python-decouple

